Theme: Fixed header FIP image's alt text.
A bilingual FIP image is situated in the top-right of all of the usability theme's pages. Until now, in all standard/server message/splash page templates, the image's alt text was unilingual and based on the current page language. However, that wasn't an accurate representation of the text in the image.

This commit adds extra handlebars conditions to make the FIP's alt text bilingual depending on the current page language. In French pages, it'll be represented as French/English bilingual text. In all other pages, it'll be represented as English/French bilingual text. Furthermore, in multilingual pages, a lang="en" attribute will be added to the image's object element in order to assist screen readers.

Related to wet-boew/GCWeb#1241.